+++ src/OpenSleigh.Core/ICompensateMessage.cs
@@ -1,0 +1,11 @@
+﻿using System.Threading;
+using System.Threading.Tasks;
+using OpenSleigh.Core.Messaging;
+
+namespace OpenSleigh.Core
+{
+    public interface ICompensateMessage<TM> where TM : IMessage
+    {
+        Task CompensateAsync(ICompensationContext<TM> context, CancellationToken cancellationToken = default);
+    }
+}