+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file.
 
+## [2021-02-24](https://github.com/mizrael/OpenSleigh/pull/27)
+### Added
+- added support for compensating transactions
+- added new sample
+
 ## [2021-02-11](https://github.com/mizrael/OpenSleigh/pull/26)
 ### Added
 - added automatic infrastructure provisioning